+++ middleware.ts
@@ -136,5 +136,6 @@ export const config = {
     '/api/roles/:path*',
     '/api/config/:path*',
     '/api/api-keys/:path*',
+    '/api/tinypng/:path*',
   ]
 }